requirements.txt
fastapi>=0.115.7
uvicorn[standard]>=0.34.2
pillow>=11.2.1
python-multipart>=0.0.20
pydantic-settings>=2.1.0
anyio>=4.8.0